requirements.txt
annotated-types==0.7.0
anyio==4.9.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
dash==3.0.0
schedule==1.2.2
dash-bootstrap-components==2.0.0
dash-html-components==2.0.0
distro==1.9.0
exceptiongroup==1.2.2
h11==0.14.0
dash-auth==2.0.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
dash-auth
MarkupSafe==3.0.2
mysql-connector-python==9.2.0
narwhals==1.32.0
nest-asyncio==1.6.0
numpy==2.2.4
openai==1.68.2
gunicorn
packaging==24.2
pandas==2.2.3
plotly==6.0.1
pydantic==2.10.6
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.3
retrying==1.3.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.39
stringcase==1.2.0
tqdm==4.67.1
tzdata==2025.2
urllib3==2.3.0
zipp==3.21.0
protobuf~=5.29.4
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
attrs==25.3.0
azure-core==1.33.0
azure-eventhub==5.14.0
azure-eventhub-checkpointstoreblob==1.2.0
azure-eventhub-checkpointstoreblob-aio==1.2.0
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.2
dnspython==2.7.0
Flask==3.0.3
frozenlist==1.5.0
idna==3.10
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
multidict==6.3.2
numpy==2.2.4
pandas==2.2.3
propcache==0.3.1
pycparser==2.22
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.3
six==1.17.0
typing_extensions==4.13.1
tzdata==2025.2
urllib3==2.3.0
yarl==1.19.0
Werkzeug==3.0.6
Flask==3.0.3
